refactor(core): necessary refactoring for the enhance command
