+++ cmd/enhance.go
@@ -2,35 +2,32 @@ package cmd
 
 import (
 	"fmt"
+	"os"
+
+	"github.com/shivanshkc/ledgerconv/core"
 
 	"github.com/spf13/cobra"
 )
 
+var enhanceParamOutputDir string
+
 // enhanceCmd represents the enhance command.
 var enhanceCmd = &cobra.Command{
-	Use:   "enhance",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Use:   "enhance <input-dir>",
+	Short: "Enhance formatted statements with custom fields.",
+	Args:  cobra.MinimumNArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("enhance called")
+		inputDir := args[0]
+		// Core call.
+		if err := core.Enhance(cmd.Context(), inputDir, enhanceParamOutputDir); err != nil {
+			_, _ = fmt.Fprintf(os.Stderr, "Failed to enhance statements: %+v\n", err)
+		}
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(enhanceCmd)
 
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// enhanceCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// enhanceCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	enhanceCmd.Flags().StringVarP(&enhanceParamOutputDir, "output", "o", "./output",
+		"The directory where the enhanced statement will be stored.")
 }
