+++ core/banks/impl_hdfc_savings.go
@@ -5,12 +5,14 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/shivanshkc/ledgerconv/core/models"
 )
 
 // convHDFCSavings converts the HDFC savings account statements to JSON.
 //
 //nolint:funlen,cyclop // Converter functions can be long.
-func convHDFCSavings(csvContent [][]string) ([]*TransactionDoc, error) {
+func convHDFCSavings(csvContent [][]string) ([]*models.TransactionDoc, error) {
 	// Bank statement CSV files do not just contain the transaction list, but also some other metadata about the
 	// bank account. This header allows us to detect the starting of the transaction table, so we can skip the needless.
 	startingHeader := []string{
@@ -20,7 +22,7 @@ func convHDFCSavings(csvContent [][]string) ([]*TransactionDoc, error) {
 	// This var will hold the index of the first transaction table row.
 	var startingIdx int
 	// This var will hold the final list of converted transactions.
-	var txDocs []*TransactionDoc //nolint:prealloc // Cannot pre-allocate this one.
+	var txDocs []*models.TransactionDoc //nolint:prealloc // Cannot pre-allocate this one.
 
 	// Trim each element. Bank statement schemas are not to be trusted!
 	for i := range csvContent {
@@ -86,7 +88,7 @@ func convHDFCSavings(csvContent [][]string) ([]*TransactionDoc, error) {
 		}
 
 		// Instantiating the transaction doc.
-		doc := &TransactionDoc{
+		doc := &models.TransactionDoc{
 			AccountName: "", // This is not the responsibility of the converterFunc.
 			Amount:      amount,
 			Timestamp:   timestamp,
