+++ core/banks/impl_icici_credit.go
@@ -5,12 +5,14 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/shivanshkc/ledgerconv/core/models"
 )
 
 // convICICICredit converts the ICICI credit card statements to JSON.
 //
 //nolint:funlen,cyclop // Converter functions can be long.
-func convICICICredit(csvContent [][]string) ([]*TransactionDoc, error) {
+func convICICICredit(csvContent [][]string) ([]*models.TransactionDoc, error) {
 	// Bank statement CSV files do not just contain the transaction list, but also some other metadata about the
 	// bank account. This header allows us to detect the starting of the transaction table, so we can skip the needless.
 	startingHeader := []string{
@@ -21,7 +23,7 @@ func convICICICredit(csvContent [][]string) ([]*TransactionDoc, error) {
 	// This var will hold the index of the first transaction table row.
 	var startingIdx int
 	// This var will hold the final list of converted transactions.
-	var txDocs []*TransactionDoc //nolint:prealloc // Cannot pre-allocate this one.
+	var txDocs []*models.TransactionDoc //nolint:prealloc // Cannot pre-allocate this one.
 
 	// Trim each element. Bank statement schemas are not to be trusted!
 	for i := range csvContent {
@@ -82,7 +84,7 @@ func convICICICredit(csvContent [][]string) ([]*TransactionDoc, error) {
 		}
 
 		// Instantiating the transaction doc.
-		doc := &TransactionDoc{
+		doc := &models.TransactionDoc{
 			AccountName: "", // This is not the responsibility of the converterFunc.
 			Amount:      amount,
 			Timestamp:   timestamp,
